traces-app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
click==8.1.8
Deprecated==1.2.18
fastapi==0.115.11
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
packaging==24.2
protobuf==5.29.4
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
zipp==3.21.0
